syscall, runtime: tv_usec is 32 bits on sparc64
SPARC is the one 64-bit Linux port whose __kernel_suseconds_t is int
rather than long - asm/posix_types.h picks it for __arch64__ as well -
so struct timeval is an 8-byte tv_sec followed by a 4-byte tv_usec and
4 bytes of padding. Both the syscall package and the runtime declared
tv_usec as int64, which on this big-endian machine reads the real
microsecond value as the high half of a 64-bit word.

The visible symptom was syscall.Getrusage: TestCPUProfileMultithreadMagnitude
compared the profile against rusage and saw 22396 hours of CPU time for
five seconds of work, then negative time once the bogus value wrapped.
Everything else built on Timeval was wrong the same way: Select,
Gettimeofday, utimes, futimesat, and the runtime's setitimer.

The vendored golang.org/x/sys/unix copy of ztypes_linux_sparc64.go
already had the right layout, which is what the two in-tree copies
should have matched.

+++ src/syscall/syscall_linux_sparc64.go
@@ -66,7 +66,7 @@ func setTimespec(sec, nsec int64) Timespec {
 }
 
 func setTimeval(sec, usec int64) Timeval {
-	return Timeval{Sec: sec, Usec: usec}
+	return Timeval{Sec: sec, Usec: int32(usec)}
 }
 
 // Time is implemented with gettimeofday: linux/sparc64 has no time(2).
